updated method and int to long

+++ app/src/main/java/cookbook/backend/be_objects/User.java
@@ -22,7 +22,7 @@ public class User {
     setPassword(password);
   }
 
-  public int getUserId() {
+  public Long getUserId() {
     return userId;
   }
   public String getUserName() {
